Fixed #14
*Fixed crash with Creeper Overhaul:
The problem was two-fold - partly due to an oversight on Data Attributes' architecture but also Creeper Overhaul.

Data Attributes gets the registry key for attributes every time an EntityAttributeInstance is constructed, and caches this so that upon reloads the game can use pass-by-data instead of pass-by-reference (i.e. attributes are now data-based instead of java object reference-based). This is a critical design choice as upon reloading attributes, the registry objects are deleted and recreated, which isn't doable with pass-by-reference. However, this relied on attributes being registered statically (as they should be, and as they are done in vanilla), as the EntityAttributeInstance's are constructed statically in each LivingEntity's createXAttributes() method.

Creeper Overhaul registers its attributes non-statically. This is not necessarily wrong, but it is weird. A solution to this bug would have been to submit a merge request and ask the author to just change registration to static method, but it's likely that other mods do this too, so it is better to just fix on my side.

The solution was simple (after actually identifying the issue): whenever an EntityAttributeInstance is called, we check if it's cached identifier is null. If it is, we reset it to the (now registered) attribute. If it's not, we don't need to do anything.

+Added modrinth buildscript to autorelease (convenient!).